changed (unsigned) to (NSUInteger)
git-svn-id: svn://orca.physics.unc.edu/Orca@11752 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Misc Objects/MemoryWatcher/MemoryWatcher.h
@@ -53,7 +53,7 @@
 - (NSTimeInterval) taskInterval;
 - (void) setTaskInterval:(NSTimeInterval)aTaskInterval;
 - (float) convertValue:(float)aValue withMultiplier:(NSString*)aMultiplier;
-- (unsigned) timeRateCount:(int)rateIndex;
+- (NSUInteger) timeRateCount:(int)rateIndex;
 - (float) timeRate:(int)rateIndex value:(int)valueIndex;
 - (void) setUpQueue;
